% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_to_img}
\alias{base64_to_img}
\title{Image data uri to file}
\usage{
base64_to_img(data_uri, slug)
}
\arguments{
\item{data_uri}{charachter, base64 image string as returned by \link{parse_search_results}}

\item{slug}{character, name of file to export image to. WITHOUT extension}
}
\value{
nothing, called for side effects
}
\description{
Convert a data uri to an image in the correct format and save it to a file.
}
\examples{
\dontrun{
data_uri <- paste0(
    "data:image/png;base64,",
    base64enc::base64encode(system.file("logo.png", package = "webbotparseR"))
)
base64_to_img(data_uri, "logo")
}
}
