% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_metadata}
\alias{parse_metadata}
\title{Parse metadata from search engine results}
\usage{
parse_metadata(path)
}
\arguments{
\item{path}{character. a path to a file that contains search results}
}
\value{
a tibble of parsed search engine results
}
\description{
Parse metadata from search engine results
}
\examples{
parse_metadata("www.google.com_climate change_text_2023-03-16_08_16_11.html")
}
