% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_search_results}
\alias{parse_search_results}
\title{Parse search engine results}
\usage{
parse_search_results(path, engine)
}
\arguments{
\item{path}{character. either a path to a file that contains search results or a path to a directory containing search engine result files}

\item{engine}{character.}
}
\value{
a tibble of parsed search engine results
}
\description{
Parse search engine results
}
